Factor out reused code for matching storage classes
